requirements.txt
python-slugify==4.0.1
selenium==3.141.0
text-unidecode==1.3
ujson==4.0.2
urllib3==1.26.4
